# backend/src/main/java/com/anveshak/DTOs/ForgotPasswordRequest.java
package com.anveshak.DTOs;

import jakarta.validation.constraints.Email;
import jakarta.validation.constraints.NotBlank;

public record ForgotPasswordRequest(
        @Email @NotBlank String email) {
}