# backend/src/main/resources/openapi.yml
openapi: 3.1.0
info:
  title: Anveshak API
  version: 1.0.0
  description: API documentation for the Anveshak research paper management application.
servers:
  - url: /
...
security:
  - bearerAuth: []
paths:
...
  /auth/register:
    post:
      tags: [Authentication]
      operationId: registerUser
      summary: Register a new user
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: Registration successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthMessageResponse'
        '400':
          description: Invalid registration request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/login:
    post:
      tags: [Authentication]
      operationId: loginUser
      summary: Login with email and password
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          description: Invalid credentials or request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/google:
    post:
      tags: [Authentication]
      operationId: googleLogin
      summary: Login with Google
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GoogleLoginRequest'
      responses:
        '200':
          description: Google login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/refresh:
    post:
      tags: [Authentication]
      operationId: refreshTokens
      summary: Refresh access and refresh tokens
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshTokenRequest'
      responses:
        '200':
          description: Tokens refreshed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          description: Invalid refresh token request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/logout:
    post:
      tags: [Authentication]
      operationId: logout
      summary: Revoke a refresh token
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshTokenRequest'
      responses:
        '204':
          description: Token revoked
        '400':
          description: Invalid logout request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/logout-all:
    post:
      tags: [Authentication]
      operationId: logoutAll
      summary: Revoke all user sessions
      responses:
        '204':
          description: All sessions revoked
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/verify-email:
    get:
      tags: [Authentication]
      operationId: verifyEmailQueryToken
      summary: Verify email using a query token
      security: []
      parameters:
        - name: token
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Email verified
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthMessageResponse'
        '400':
          description: Invalid verification token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/verify-email/{token}:
    get:
      tags: [Authentication]
      operationId: verifyEmailPathToken
      summary: Verify email using a path token
      security: []
      parameters:
        - name: token
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Email verified
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthMessageResponse'
        '400':
          description: Invalid verification token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/resend-verification:
    post:
      tags: [Authentication]
      operationId: resendVerification
      summary: Resend verification email
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResendVerificationTokenRequest'
      responses:
        '200':
          description: Verification email sent
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthMessageResponse'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/forgot-password:
    post:
      tags: [Authentication]
      operationId: forgotPassword
      summary: Request a password reset email
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ForgotPasswordRequest'
      responses:
        '200':
          description: Password reset email sent
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthMessageResponse'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/reset-password:
    post:
      tags: [Authentication]
      operationId: resetPassword
      summary: Reset a password using a reset token
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResetPasswordRequest'
      responses:
        '200':
          description: Password reset successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthMessageResponse'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /papers/{paperId}/citations:
    post:
      tags: [Citation Management]
      operationId: createCitation
      summary: Create a citation
      parameters:
        - name: paperId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewCitationRequest'
      responses:
        '201':
          description: Citation created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CitationResponse'
    get:
      tags: [Citation Management]
      operationId: listCitations
      summary: List citations for a paper
      parameters:
        - name: paperId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Citations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CitationResponse'
  /papers/{paperId}/citations/{citationId}:
    delete:
      tags: [Citation Management]
      operationId: deleteCitation
      summary: Delete a citation
      parameters:
        - name: paperId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: citationId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Citation deleted
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: jwt
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        message:
          type: string
        code:
          type: integer
    AuthMessageResponse:
      type: object
      required: [message]
      properties:
        message:
          type: string
    AuthResponse:
      type: object
      required: [accessToken, refreshToken]
      properties:
        accessToken:
          type: string
        refreshToken:
          type: string
    RegisterRequest:
      type: object
      required: [firstName, lastName, username, email, password]
      properties:
        firstName:
          type: string
        lastName:
          type: string
        username:
          type: string
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
    GoogleLoginRequest:
      type: object
      properties:
        token_id:
          type: string
          description: Google ID token.
    RefreshTokenRequest:
      type: object
      required: [token]
      properties:
        token:
          type: string
        id:
          type: string
          nullable: true
          description: Optional refresh-token identifier.
    ForgotPasswordRequest:
      type: object
      required: [email]
      properties:
        email:
          type: string
          format: email
    ResetPasswordRequest:
      type: object
      required: [token, newPassword]
      properties:
        token:
          type: string
        newPassword:
          type: string
    ResendVerificationTokenRequest:
      type: object
      required: [email]
      properties:
        email:
          type: string
          format: email
...
    NewCitationRequest:
      type: object
      required: [citedPaperId]
      properties:
        citedPaperId:
          type: string
          format: uuid
    CitationResponse:
      type: object
      required: [id, citingPaperId, citingPaperTitle, citedPaperId, citedPaperTitle, createdAt]
      properties:
        id:
          type: string
          format: uuid
        citingPaperId:
          type: string
          format: uuid
        citingPaperTitle:
          type: string
        citedPaperId:
          type: string
          format: uuid
        citedPaperTitle:
          type: string
        createdAt:
          type: string
          format: date-time
...